app/requirements.txt
Flask==2.2.2
tensorflow==2.11.1
Pillow==9.2.0